+++ lib/valvat/lookup.rb
@@ -8,11 +8,15 @@ module Valvat
       parts = Valvat::Utils.split(vat)
       return false unless parts[0]
       
-      YAML.load(Net::HTTP.start("isvat.appspot.com", 80) {|http|
-        http.get("/#{parts.join("/")}/")
-      }.body)
-    rescue
-      nil
+      result = begin
+        YAML.load(Net::HTTP.start("isvat.appspot.com", 80) {|http|
+          http.get("/#{parts.join("/")}/")
+        }.body)
+      rescue
+        nil
+      end
+      
+      result.is_a?(Hash) && result["error_code"] == 1 ? nil : result
     end
   end
 end